refactor: created ipxact_builder and migrated IPXACT2022Generator to there
