+++ tools/utils/__init__.py
@@ -1,1 +1,2 @@
 from .latex_parser import *
+from .ipxact_builder import IPXACT2022Generator
